+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## CHANGELOG ##
 
+* @dev
+  * `name()` argument is now `trim()`ed automatically.
+
 * v1.3.0 (2020-09-23)
   * Bumped min PHP version to 5.6
   * Added Travis-CI support.