+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * @dev
   * `name()` argument is now `trim()`ed automatically.
+  * Removed `invobill()` method. Use `reserved2()` instead.
 
 * v1.3.0 (2020-09-23)
   * Bumped min PHP version to 5.6